# techarrow_mobile_final/api_files/swagger.json
{"openapi":"3.1.0","info":{"title":"tetris","license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"paths":{"/api/advice/create":{"post":{"tags":["Advice"],"summary":"Create Advice","operationId":"create_advice_api_advice_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdviceCreateRq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdviceCreateRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/advice/get":{"get":{"tags":["Advice"],"summary":"Get Advice","operationId":"get_advice_api_advice_get_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdviceGetRs"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/fields/update":{"put":{"tags":["Fields"],"summary":"Update Field","operationId":"update_field_api_fields_update_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldAnswer"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/fields/get/{date}":{"get":{"tags":["Fields"],"summary":"Get Field","operationId":"get_field_api_fields_get__date__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"date","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FieldOutput"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ping":{"get":{"tags":["Healthcheck"],"summary":"Ping","operationId":"ping_api_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthcheckPingRs"}}}}}}},"/api/tasks/create":{"post":{"tags":["Tasks"],"summary":"Create Task","operationId":"create_task_api_tasks_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreateRq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreateRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/tasks/to_planned":{"put":{"tags":["Tasks"],"summary":"To Planned","operationId":"to_planned_api_tasks_to_planned_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToPlanRq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToPlanRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/tasks/get/{data}":{"get":{"tags":["Tasks"],"summary":"Get By Date","operationId":"get_by_date_api_tasks_get__data__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"data","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Data"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListDateRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/to_completed/{task_id}":{"put":{"tags":["Tasks"],"summary":"To Completed","operationId":"to_completed_api_tasks_to_completed__task_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToPlanRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/delete/{task_id}":{"delete":{"tags":["Tasks"],"summary":"Delete Task","operationId":"delete_task_api_tasks_delete__task_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToPlanRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/get_unplanned":{"get":{"tags":["Tasks"],"summary":"Get Unplanned","operationId":"get_unplanned_api_tasks_get_unplanned_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListRs"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/tasks/get_planned":{"get":{"tags":["Tasks"],"summary":"Get Planned","operationId":"get_planned_api_tasks_get_planned_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListRs"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/tasks/get_completed":{"get":{"tags":["Tasks"],"summary":"Get Completed","operationId":"get_completed_api_tasks_get_completed_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListRs"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/tasks/get_by_id/{task_id}":{"get":{"tags":["Tasks"],"summary":"Get By Id","operationId":"get_by_id_api_tasks_get_by_id__task_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskGetRs"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdviceCreateRq":{"properties":{"ids":{"items":{"type":"integer"},"type":"array","title":"Ids"}},"type":"object","required":["ids"],"title":"AdviceCreateRq"},"AdviceCreateRs":{"properties":{"id":{"type":"integer","title":"Id"},"text":{"type":"string","title":"Text"},"color":{"type":"string","title":"Color"}},"type":"object","required":["id","text","color"],"title":"AdviceCreateRs"},"AdviceGetRs":{"properties":{"advice":{"items":{"$ref":"#/components/schemas/AdviceCreateRs"},"type":"array","title":"Advice"}},"type":"object","required":["advice"],"title":"AdviceGetRs"},"FieldAnswer":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"FieldAnswer"},"FieldInput":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"list":{"items":{"items":{"type":"integer"},"type":"array"},"type":"array","title":"List"}},"type":"object","required":["date","list"],"title":"FieldInput"},"FieldOutput":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"matrix":{"items":{"items":{"type":"integer"},"type":"array"},"type":"array","title":"Matrix"},"info":{"items":{"$ref":"#/components/schemas/TaskInfo"},"type":"array","title":"Info"}},"type":"object","required":["date","matrix","info"],"title":"FieldOutput"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthcheckPingRs":{"properties":{"status":{"type":"string","title":"Status"},"response_time":{"type":"number","title":"Response Time"}},"type":"object","required":["status","response_time"],"title":"HealthcheckPingRs"},"SubTask":{"properties":{"name":{"type":"string","title":"Name"},"duration":{"type":"string","title":"Duration"}},"type":"object","required":["name","duration"],"title":"SubTask"},"TaskCategory":{"type":"string","enum":["learning","work","personal"],"title":"TaskCategory"},"TaskCreateRq":{"properties":{"type":{"$ref":"#/components/schemas/Type"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"duration":{"type":"string","title":"Duration"},"importance":{"type":"boolean","title":"Importance"},"category":{"$ref":"#/components/schemas/TaskCategory"},"subtasks":{"items":{"$ref":"#/components/schemas/SubTask"},"type":"array","title":"Subtasks"}},"type":"object","required":["type","name","description","importance","category"],"title":"TaskCreateRq"},"TaskCreateRs":{"properties":{"task_id":{"type":"integer","title":"Task Id"}},"type":"object","required":["task_id"],"title":"TaskCreateRs"},"TaskGetRs":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"category":{"$ref":"#/components/schemas/TaskCategory"},"duration":{"type":"string","title":"Duration"},"importance":{"type":"boolean","title":"Importance"},"is_multi":{"type":"boolean","title":"Is Multi"},"multi_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Id"},"multi_color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Color"},"multi_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Name"},"multi_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Description"}},"type":"object","required":["id","name","description","category","duration","importance","is_multi","multi_id","multi_color","multi_name","multi_description"],"title":"TaskGetRs"},"TaskInfo":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"category":{"$ref":"#/components/schemas/TaskCategory"},"duration":{"type":"string","title":"Duration"},"importance":{"type":"boolean","title":"Importance"},"is_multi":{"type":"boolean","title":"Is Multi"},"multi_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Id"},"multi_color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Color"},"multi_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Name"},"multi_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Multi Description"}},"type":"object","required":["id","name","description","category","duration","importance","is_multi","multi_id","multi_color","multi_name","multi_description"],"title":"TaskInfo"},"TaskList":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"duration":{"type":"string","title":"Duration"}},"type":"object","required":["id","name","description","duration"],"title":"TaskList"},"TaskListDate":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"TaskListDate"},"TaskListDateRs":{"properties":{"planned_tasks":{"items":{"$ref":"#/components/schemas/TaskListDate"},"type":"array","title":"Planned Tasks"},"completed_tasks":{"items":{"$ref":"#/components/schemas/TaskListDate"},"type":"array","title":"Completed Tasks"}},"type":"object","required":["planned_tasks","completed_tasks"],"title":"TaskListDateRs"},"TaskListRs":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/TaskList"},"type":"array","title":"Tasks"}},"type":"object","required":["tasks"],"title":"TaskListRs"},"TaskToPlanRq":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"tasks":{"items":{"type":"integer"},"type":"array","title":"Tasks"}},"type":"object","required":["date","tasks"],"title":"TaskToPlanRq"},"TaskToPlanRs":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"TaskToPlanRs"},"Type":{"type":"string","enum":["unplanned","multifaceted"],"title":"Type"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}